fixes a few broken rollbacks in migration. also adds indexes on other…

+++ database/migrations/2024_01_30_163819_add_id_locked_field_to_dur_has_datasets_table.php
@@ -20,9 +20,11 @@ return new class () extends Migration {
      */
     public function down(): void
     {
-        Schema::table('dur_has_datasets', function (Blueprint $table) {
-            $table->dropColumn('is_locked');
-        });
+        if (Schema::hasTable('dur_has_datasets')) {
+            Schema::table('dur_has_datasets', function (Blueprint $table) {
+                $table->dropColumn('is_locked');
+            });
+        }
     }
 };
